-> Refactoring of OffHeap version

+++ core/src/main/java/org/mwdb/chunk/KChunk.java
@@ -10,7 +10,7 @@ public interface KChunk {
 
     byte chunkType();
 
-    int marks();
+    long marks();
 
     long flags();
